# specs/002-notification-system/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: UniTeam Notification API
  version: 1.0.0
  description: 알림 시스템 REST 계약 (002-notification-system)
servers:
  - url: https://p18.sumzip.com/api/v1
    description: Production
  - url: http://localhost:9538/api/v1
    description: Local backend
security:
  - bearerAuth: []
paths:
  /notifications:
    get:
      summary: 알림 목록 조회
      parameters:
        - name: status
          in: query
          schema: { type: string, enum: [unread, read, archived, all], default: all }
        - name: cursor
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 50, default: 20 }
      responses:
        '200':
          description: 알림 목록
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items: { $ref: '#/components/schemas/Notification' }
                  nextCursor: { type: string, nullable: true }
                  unreadCount: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /notifications/unread-count:
    get:
      summary: 안읽음 개수(폴링용)
      responses:
        '200':
          description: 안읽음 개수
          content:
            application/json:
              schema:
                type: object
                properties:
                  unreadCount: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /notifications/{id}/read:
    post:
      summary: 개별 알림 읽음 처리(멱등)
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        '200':
          description: 처리된 알림
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Notification' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
  /notifications/read-all:
    post:
      summary: 전체 읽음 처리
      responses:
        '200':
          description: 처리 결과
          content:
            application/json:
              schema:
                type: object
                properties:
                  updated: { type: integer }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /notifications/preferences:
    get:
      summary: 수신 설정 조회
      responses:
        '200':
          description: 종류별 채널 설정 + 전역 설정
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PreferenceBundle' }
        '401': { $ref: '#/components/responses/Unauthorized' }
    put:
      summary: 수신 설정 갱신
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/PreferenceBundle' }
      responses:
        '200':
          description: 갱신된 설정
          content:
            application/json:
              schema: { $ref: '#/components/schemas/PreferenceBundle' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Notification:
      type: object
      required: [id, type, priority, title, body, status, createdAt]
      properties:
        id: { type: string }
        type: { type: string }
        priority: { type: string, enum: [critical, normal, info] }
        title: { type: string }
        body: { type: string }
        deepLink: { type: string, nullable: true }
        groupCount: { type: integer }
        status: { type: string, enum: [unread, read, archived] }
        createdAt: { type: string, format: date-time }
        readAt: { type: string, format: date-time, nullable: true }
    ChannelPreference:
      type: object
      required: [type, inApp, email, push, mandatory]
      properties:
        type: { type: string }
        inApp: { type: boolean }
        email: { type: boolean }
        push: { type: boolean }
        mandatory: { type: boolean, description: true면 비활성 불가 }
    GlobalSettings:
      type: object
      properties:
        dndEnabled: { type: boolean }
        quietStart: { type: string, example: "22:00" }
        quietEnd: { type: string, example: "08:00" }
        timezone: { type: string, example: "Asia/Seoul" }
    PreferenceBundle:
      type: object
      properties:
        preferences:
          type: array
          items: { $ref: '#/components/schemas/ChannelPreference' }
        global: { $ref: '#/components/schemas/GlobalSettings' }
    Error:
      type: object
      properties:
        requestId: { type: string }
        code: { type: string }
        message: { type: string }
  responses:
    Unauthorized:
      description: 인증 필요
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: 대상 없음(EC-06)
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    BadRequest:
      description: 잘못된 요청
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }